+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add CA certificates to Docker image for AWS S3 access, [PR-924](https://github.com/reductstore/reductstore/pull/924)
 - Fix path for token repository when cache used, [PR-925](https://github.com/reductstore/reductstore/pull/925)
+- Prevent storage from crashing if block from WAL doesn't exist, [PR-954](https://github.com/reductstore/reductstore/pull/954)
 
 ### Internal